Native M1 and Codesigning Fixes
Unfortunately, cross compiling will result in codesigning error for the
non-native arch of the build system. Still, it's an improvement as the
the app was not configured to build for native ARM and users of x86_64
systems ought to have no problem compiling. Plus, codesigning and
notarization is mandatory for M1 Macs by default.